python: Fix test 505
The logic in this test was wrong, but masked by other problems still
existing in our python bindings until recently.  The fourth test
claims to fail during only chunk, but passed '42' as the user_data for
callback which requires that err.value be 0; if .chunk had actually
failed, err.value should be non-zero.

We're overloading a single user_data int to try and control two
orthogonal things (what to expect in err, and whether to fail); let's
instead make the callback user_data a tuple.  Counterpart to ocaml
commit 92bd7b08.

Fixes: e7d29712