ocaml: Fix test 505
The logic in this test was wrong (blindly copied from a similar test
in python), but masked by other problems until recently.  The fourth
test claims to fail during only chunk, but passed '42' as the
user_data for callback which requires that !err be 0; if chunk had
actually failed, !err should be non-zero.

We're overloading a single user_data int to try and control two
orthogonal things (what to expect in err, and whether to fail); let's
instead make the callback user_data a tuple.

Fixes: ce0f2126