ocaml: Use caml_alloc_custom_mem, if available, to hint off-heap size.
This function allows custom blocks to tell the garbage collector how
much off-heap (ie. not in OCaml heap but in C heap) memory they
represent, allowing the GC to make better decisions.  For persistent
buffers we have a very good idea of this.  For NBD handles, less so,
but we can still do better than we are currently.

Note this function is relatively recent, and replaces an older
mechanism which was much harder to use correctly, and we didn't really
use it properly, just passing (0, 1) as parameters.