generator: Add wrapping mode.
This commit modifies the pr function to allow for wrapped text.  To do
this use (eg to wrap on commas):

  pr_wrap ',' (fun () ->
                 (* some code using pr *) )

The wrapping takes effect only for the duration of the inner function.
The current output column is saved and further lines are indented by
the same amount.  For example if pr_wrap is called immediately after
the '(' is printed then it wraps as follows:

  int64_t nbd_aio_pread_structured (struct nbd_handle *h, void *buf,
                                    size_t count, uint64_t offset,
                                    nbd_chunk_callback chunk_callback,
                                    nbd_completion_callback completion_callback,
                                    uint32_t flags);
                                    ^
                                    |
                   saved wrapping column

You can wrap on any character, but only ',' has been tested.