docs: Include examples in generated man pages.
Allow examples to be included in generated man pages:

  "connect_unix", {
    default_call with
    ...
    example = Some "examples/fetch-first-sector.c";
  };

The examples come out of the examples/ subdirectory, and one advantage
of this is that they are compiled during the build process so we know
that they at least build correctly.

Example coverage is fairly poor at the moment but that can be improved
later.