docs: Fix option styling.
This commit fixes the man pages in a few places to be consistent with
the usual style, described below.

(a) If it's something you would type exactly, it's all in bold, eg:

    =item B<cache=unsafe>

because “cache=unsafe” actually appears on the command line.

(b) Otherwise the part you would type is in bold, and the rest is
usually¹ capitalized, eg:

    =item B<cache-max-size=>SIZE
    =item B<rdelay=>NNB<ms>

because “cache-max-size=1M” or “rdelay=30ms” is what would appear on
the command line.

(c) Alternatives are written on separate lines instead of using ‘|’,
eg:

    =item B<fuamode=none>

    =item B<fuamode=emulate>

Note that a blank line between alternatives is required by POD, but
POD also treats them as effectively a single heading/bullet.

¹ We're not very consistent about capitalization.

(cherry picked from nbdkit commit f2087dc03b12ef0874ec8a31bbc5351a3c262b37)