configure: Ban use of Variable Length Arrays (VLAs).
Since we don't know much about the calling environment, which might
have a limited stack, might be taking input from untrusted sources, or
we might not have other stack protections enabled, it's best to be
cautious about using unbounded stack allocations.

We're not in fact using them in libnbd, but this change prevents them
from being added in future.