nbdsh: Add --base-allocation option to simplify h.block_status
When we first added nbd_block_status, we decided to not request the
"base:allocation" context by default (if a client wants to use
block_status on a different context, then they'd have to get any
default request out of the way); however, block status is useless
without at least one meta context.  At the same time, you can't
request block status after connection, which means if you start nbdsh
with --connect, it's too late.  So this adds a convenience knob for
requesting "base:allocation" with less typing.  Note that '--b' is
recognized as a synonym for the longer --base-allocation (as long as
we don't introduce any other long-option starting with --b).

I specifically chose not to burn a short option. After all, '-b' for
'base-allocation' might instead be useful for 'buffer' or 'blocksize'
down the road, '-c' for context is already ambiguous with --connect
and --command, '-a' for allocation might clash with 'append', and so
on.  What's more, we may decide to change the default to always
requesting "base:allocation", coupled with an API to allow a client to
clear that default (choosing whether their alternative context is done
in addition or in place of the default); it would mean a slightly
longer handshake for clients that don't care about calling
nbd_block_status, but less hassle and no need for the command line
option added here if we set up that default.