nbdsh: Prefer --uri over --connect
Typing nbdsh --con (the minimum to get an unambiguous prefix for
--connect, different from --command) is annoying compared to having a
short option.  Since it takes a URI as an argument, using -u/--uri is
a nicer mnemonic.  We still accept --connect for back-compat, and
document it in the man page, but omit it from --help as the new
spelling is nicer all around.