interop: Use systemd socket activation with qemu-nbd.
Now that we have implemented systemd socket activation we can use this
as a way to test qemu-nbd instead of having to use a TCP port.

After this change there are no servers which require a TCP port for
interop — all of nbdkit, qemu-nbd and nbd-server use either stdin or
socket activation — and so the -DSERVE_OVER_TCP=1 code can be removed
completely which considerably simplifies interop.c.

Thanks: Eric Blake.