docs: Add libnbd-security(3) man page
Copies heavily after a similar addition recently made in nbdkit.

Placed in section 3 since the rest of the library man pages live
there.  In nbdkit, it was in section 1, but while the nbdkit project
has the 'nbdkit(1)' binary, we only ship the 'nbdsh(1)' binary, but
the security information is for the project not just the binary.