api: Add nbd_connect_socket.
This allows us to connect directly to a connected socket.  How exactly
the socket is created and connected to the NBD server is left up to
the main program.

The only real complexity in this patch is allowing file descriptors to
be passed to libnbd APIs.  Luckily in C and Python we can treat them
exactly as integers, and in OCaml they are also integers but with a
different type.