nbdfuse: New tool to present a network block device in a FUSE filesystem.
This program allows you to turn a network block device source into a
FUSE filesystem containing a virtual file:

  $ nbdkit memory 128M
  $ mkdir mp
  $ nbdfuse mp/ramdisk nbd://localhost &
  $ ls -l mp
  total 0
  -rw-rw-rw-. 1 rjones rjones 134217728 Oct 12 15:09 ramdisk
  $ dd if=/dev/urandom bs=1M count=128 of=mp/ramdisk conv=notrunc,nocreat
  128+0 records in
  128+0 records out
  134217728 bytes (134 MB, 128 MiB) copied, 3.10171 s, 43.3 MB/s
  $ fusermount -u mp

There are still some shortcomings, such as lack of zero and trim
support.  These are documented in the TODO file.

Now for some history:

In libguestfs which is where most of this code derives from we have a
program called ‘guestmount’ which is a FUSE interface to libguestfs:

  http://libguestfs.org/guestmount.1.html

Originally that was a standalone program like nbdfuse, but after some
time we realized that the ability to mount libguestfs under a
directory was generally useful to all guestfs API users and we created
new APIs for it.  guestmount has now become a thin wrapper around
those APIs.

This of course argues that we should do the same thing for libnbd.
But ...

(1) For NBD this is a little less useful than for libguestfs.

(2) We can always do this in future if we need to.

Most importantly:

(3) the libguestfs FUSE API turned out to have a problem - still
unresolved - with handling threads and SELinux and it may not be a
good idea to bake this into the libnbd API until that problem has been
solved.  For more details about (3), read this:

  https://bugzilla.redhat.com/show_bug.cgi?id=1060423#c2