api: Add support for AF_VSOCK.
This adds a new API for connecting to AF_VSOCK protocol
(https://wiki.qemu.org/Features/VirtioVsock).

For example:

  nbd_connect_vsock (nbd, 2, 10809);

There is no test of this feature because it only works between guest
and host.  You cannot start a server and client on the host and talk
between them, which is what we'd need to write a sane test.