docs: Set SYNOPSIS section of man pages in verbatim.
Previously we used a strange mix, especially when long API calls were
wrapped.  The generated POD looked like:

int nbd_set_debug_callback (struct nbd_handle *h,
                            nbd_debug_callback debug_callback);

Because of the way that POD works this means the first line is
non-verbatim and the second line is verbatim.  The generated man page
looks fine but the HTML is weird.

With this change all parts of the SYNOPSIS section have a space
prefixed so are set in verbatim.