generator: pr_wrap ?maxcol parameter to control max wrap column.
When wrapping long lines, we formally wrapped if the line would go
beyond column 76.  However when typesetting verbatim text in man
pages, this is too long and a shorter maximum line length would be
preferable.  This new optional parameter allows this to be chosen.

Defaults to 76, so there is no change to the output in this commit.

The print_arg_list, print_cbarg_list and print_call functions have
also been updated to take this optional parameter.