generator: Split the generator script into separate files and compile it.
The generator was over 6000 lines of code.  Split it into separate
files to make it more manageable.  This also forces the interfaces
between parts of the generator to be properly documented (ie. the
*.mli files).

Instead of interpreting it as a big script, we compile it using the
OCaml bytecode compiler (ocamlc).  The option of continuing to
interpret the script and using #use for textual inclusion was
attempted but in the end did not work well
(https://sympa.inria.fr/sympa/arc/caml-list/2020-02/msg00051.html)

We now require the bytecode compiler instead of the OCaml interpreter.
The bytecode compiler is available on all platforms, which is better
than requiring the native OCaml compiler which is not available
everywhere.  In practical terms this means we require ocamlc (instead
of ocaml) to build from git.  In Fedora ocaml and ocamlc are packaged
together.  In Debian they are packaged separately (ocaml-interp vs
ocaml-nox).

This is refactoring.  The generated output does not change.