build: Allow C programs using libnbd to be compiled against build dir.
We use a similar trick to libvirt and libguestfs to allow external C
programs that use libnbd to be compiled against the built (but not
installed) libnbd with:

  ../libnbd/run ./configure
  make

What actually happens is we have a second pkg-config file
(lib/local/libnbd.pc) which points to the locally built libnbd.  The
./run script sets up PKG_CONFIG_PATH to point to this directory.
Assuming that ./configure is using pkg-config/pkgconf and not some
other half-baked solution it will pick up the libnbd.pc file from here
which will set CFLAGS and LIBS appropriately.

See also this libguestfs commit:
https://github.com/libguestfs/libguestfs/commit/e33b3c83a02c1d7e82018105166e451373e8c218