Add Go language bindings (golang) (RHBZ#1814538).
This adds a mostly complete implementation of libnbd bindings for
golang.  Synchronous calls and callbacks are fully supported.  There
is a reasonably complete test suite.

Asynchronous calls are also working, with the caveat that the AIO
buffer is not GC-safe (as in other languages).  At the moment the
caller has to ensure that the AIO buffer passed in when starting the
command remains live until the command has completed (like in C).  We
may fix this properly in future, hopefully without needing to change
the golang API.

Thanks: Daniel P. Berrangé for invaluable help with this feature.