generator: Tweak internal cross-referencing comments
While working on a potential future extension, I found it awkward that
our code for enum values is rather distant from our code for #define'd
constants.  Although C code built against the resulting libnbd.h
doesn't care which of the two was used (they both serve as
compile-time constants), it is a bit harder when adding something to
the API to decide where to add it if you can't quickly locate the
correct section of the generator where a new value should be added.