states: Add state for shutdown/gnutls_bye after NBD_CMD_DISC
nbdkit prior to commit c70616f87c was prone to deadlock on a
bi-directional connection: it would end up stuck in a read() after
getting NBD_CMD_DISC to make sure that the client wasn't erroneously
sending any more requests, at the same time that the client could get
stuck in a read() waiting for EOF from the server responding to
NBD_CMD_DISC.  While nbdkit as a server has been fixed, it is possible
that there are other servers out there with the same issue.  What's
more, nbdkit 430f814102 demonstrated that clients can be robust to
such servers, by utilizing shutdown() to give the server an early EOF
even while the bi-directional socket remains open.  So, libnbd needs
to use the same solution as what nbdkit had done as a client.

We need to wire up a couple of new states in order to call the
just-added .shut_writes hook as needed (in practice, a client won't be
calling nbd_aio_disconnect while there are still pending earlier
requests, and thus we are unlikely to hit the case where writes
performed during gnutls_bye() actually block; but the logic to support
this is easy to copy from what we already have in place for
NBD_CMD_WRITE).