sockets: Add .shut_writes callback
The next commit wants to teach libnbd to ensure that a server will
read EOF from a bi-directional connection as soon as libnbd is done
writing, rather than making libnbd dependent on the server being the
connection endpoint to initiate closing the socket or waiting on the
user to eventually call nbd_close().  For that to work, we need to
expose the equivalent of shutdown(SHUT_WR) for our various connection
types.

Note that shutdown() is only applicable to sockets, since it only
really aids in bi-directional communication on a single fd.  For now,
libnbd only uses sockets (even for nbd_connect_command, we use
socketpair() to pass a single fd to both stdin and stdout, rather than
two pipe() calls to create a pair of one-way communication).  But if
we were to ever support NBD connections over a pair of pipe()s, the
obvious implementation would be to close() the write pipe during
.shut_writes, while leaving the read pipe open until .close.

Also note that shutdown() is effectively non-blocking, but that
gnutls_bye() can fail with GNUTLS_E_AGAIN if it is waiting for space
to send notification to the server.  But at the same time, while
socket read and write normally distinguish between success, EAGAIN,
and other failures, we don't need the tri-state return from our new
hook (we will only ever attempt to shut down writes after sending
NBD_CMD_DISC, and moving to CLOSED instead of DEAD at that point
doesn't change the results of any earlier requests).

Finally, note that the semantics of this new callback are specifically
only for SHUT_WR, and not generic enough for SHUT_RDWR.  This is
because gnutls_bye() handles SHUT_WR with only writes to the socket,
but when handling SHUT_RDWR it also waits for a subsequent read.
Getting the ping-pong I/O correct would require more states, for no
real gain when the user can just blindly call nbd_close() and close
the entire socket rather than having to massage their poll() loop
through more states.