golang: Allow bindings to be compiled against different libnbd.h.
Because it appears that we will need to ship the golang bindings
separately, allow them to be compiled standalone, possibly against
older or newer libnbd.h.

There are some compromises we have to make: We cannot use <config.h>
(since that is generated by the libnbd build).  We want to use
asprintf so we must manually define _GNU_SOURCE.  Any other features
of <config.h> cannot be used, but luckily we don't use them.

I tested this with libnbd 1.2.2 and it worked.