ocaml: Fix bug in asynch copy test.
This program had a subtle bug: If we've read everything from the
source, but there are still source commands in flight, AND it happens
that there are no destination commands in flight, then the loop
finishes prematurely.  This is particularly evident when you run the
test on a slow machine (armv7 in Koji is illustrative).  The fix is to
check that there are no commands in flight either from the source or
to the destination.

(I actually discovered this when writing the golang bindings which
copied from the OCaml version, but forgot to make the correction to
the original.)

(cherry picked from commit 0475bfe04a527051c0a37af59a733c4c8554e427)