New nbdcopy tool.
This tool is currently very simplistic.  For example it makes no
attempt to be efficient or keep multiple requests in flight, and there
are other shortcomings (see TODO).  It's designed as a placeholder for
future work and to allow us to stream in and out from stdin/stdout,
which is something that cannot currently be done with qemu-img
convert, but is very desirable for KubeVirt.