info: Add nbdinfo -L feature to list all exports.
In this mode, content probing is disabled by default.  This is to
allow for a future optimization where instead of doing multiple
connections to fetch the metadata for each export we could do only a
single connection with multiple NBD_OPT_INFO requests.  Probing
content would still require multiple connections.