nbdinfo: Quote use of $uri in nbdkit --run
Since a Unix socket URI includes a bare ?, it is subject to unwanted
shell globbing in the (admittedly unlikely) case the user has an
'nbd+unix:' subdirectory in their tree.  Set a good example by quoting
things.  $VG is exported by our 'run' script, so it doesn't matter if
we defer its expansion to --run time.

Signed-off-by: Eric Blake <eblake@redhat.com>