states: Request NBD_INFO_BLOCK_SIZE during NBD_OPT_GO
The NBD spec allows servers to refuse NBD_OPT_GO to clients that do
not promise to obey block size constraints (see
NBD_REP_ERR_BLOCK_SIZE_REQD).  Upcoming patches will make libnbd able
to remember and honor block sizes, but even when we don't honor them,
generally the worst that will happen is the server will send us EINVAL
errors on unaligned I/O, which is better than complete failure to
connect by not even asking about block sizes.  To ask, we only need to
add NBD_INFO_BLOCK_SIZE as a client info in our NBD_OPT_GO.  The NBD
spec is clear that servers that don't understand client requests
should silently ignore the request, so it doesn't hurt to always ask
(but if we ever _do_ find a non-compliant server where it matters, or
if we wanted to improve unit testing of a server's ability to send the
REP_ERR_BLOCK_SIZE_REQD error, we could add a knob to tell libnbd to
skip the ask for such servers).  For now, we just stash what we learn
(if anything) in internal state; the next patch will then add API to
expose it to the end user.