api: Add new API to query negotiated block size
Servers such as qemu can already tell us their preferred block size,
and work is underway to let nbdkit do likewise.  As this information
can prove useful to clients, we need a new API to expose it.  The API
is designed to support future enum values if the NBD protocol is ever
extended further to advertise larger limits for trim, zero, and cache
operations.