api: Get description alongside name from NBD_REP_SERVER
At least 'qemu-nbd -D' can send a description when listing an export;
it's not that much harder to expose things via a new
nbd_get_list_export_description.  We may still want to rethink our
list mode API (perhaps adding a callback that informs the client of
name/description pairs as they come in, instead of malloc'ing a copy
for browsing later), but for this patch, it is a straightforward
expansion building on top of commit c2851cf5.