api: Permit export list APIs when Connected
Our nbd_can_FOO APIs are allowed when Connected and Closed, but not
when Dead (which makes sense: they tell what the server answered while
connecting, and are valid whether we are still talking to the server
or cleanly disconnected, but if we are in Dead, we cannot tell if the
server answered us).  But our recent addition of list export APIs were
originally only allowed when Closed or Dead.  Dead makes sense here
(that is our default state if NBD_OPT_GO fails, but failure is likely
if the server did not like our export name; but the point of listing
was to figure out what export name to pass), but we also need to
permit Connected (because the list is accurate even if we did manage
to successfully connect).

The testsuite changes demonstrate why this change is needed (changing
just the tests without the generator causes the tests to fail), but
also highlights another issue: we are holding open two connections at
once to the same server, which deadlocks if the server in question
only permits one client at a time.