api: Add nbd_is_state_negotiating for new state
This patch prepares for a change to export list extraction by adding a
new state in the state machine, NEGOTIATING.  At present, it is not
actually possible to enter this state.  But upcoming patches will
allow the user to opt-in to fine-grained control over negotiation
during the connecting phase.  Since the handshake phase is synchronous
(the client sends a full command, then waits for the full server
response, with no interleaving or out-of-order traffic), the idea is
that we can drop out of connecting over to Negotiating any time we are
waiting for the user to determine the next client command to send, and
then back to connecting until that option is handled; where unlike the
data phase, we won't need a cookie.  Other upcoming patches will then
adjust existing API that make sense to call while in this state, as
well as adding new API to give the user control over when to actually
use this state.