api: Add nbd_set_opt_mode
Now that we have a new NEGOTIATING state, we need a knob for the user
to request moving into that state.  This patch does not actually alter
the state machine, but one thing at a time.  However, the libnbd docs
now mention several commands that will be added in future commits.
(Hmm - our generator refuses to compile in-tree links in API.ml that
don't resolve, but not links in libnbd.pod.)

One thing we can test at this point: an oldstyle server will never
get into opt mode; that will only be possible for a newstyle server.