api: add nbd_opt_info, nbd_aio_opt_info
Make it possible to query details about a potential export without
actually connecting.  This reuses the NBD_OPT_GO state machine, with
the main difference being that we can now return to negotiating state
even on success; thus checking for progression to Ready state is no
longer sufficient for detecting failure, and we have to rely on the
completion callback.

And now that we can get to Negotiation after a successful query, we
have to enable a lot more functions for use in that state to read the
results of the info query.

The next patch will then update nbdinfo to take advantage of this.
Meanwhile, this adds testsuite coverage of all our language bindings
(while the test is transcribed to be the same between them, I'm less
certain my ocaml and golang versions are idiomatic).