info: Improve info without --list
Make use of the new nbd_get_export_description to list a description
when connecting to a single export.

At present, only qemu-nbd offers the needed information; it would
require a patch to nbdkit to do likewise (although such a patch is not
difficult, given that we recently added .list_exports).