states: Fix bug in NBD_OPT_EXPORT_NAME with non-empty export
How embarrassing: on little-endian machines, a 0-byte export name
works, but a 5-byte export name is enough to make nbdkit reject the
client out-right, and a 1-byte export name causes deadlock where the
server is stuck waiting for a 16M option that will never arrive, and
the client is waiting for a response from the server.

A quick search of 'grep "sbuf.*= [^h]' did not find any other endian
bugs of a similar nature.

Add testsuite coverage, copying liberally from the existing oldstyle
test; the lack of coverage explains why we failed to detect this for
so long (although admittedly, 'nbdkit --mask-handshake' did not exist
when libnbd first added NBD_OPT_EXPORT_NAME support).

Fixes: 8b8ed26c71