api: Add nbd_opt_list
Adding NBD_OPT_LIST support will occur across two commits.  This patch
switches the existing nbd_set_list_exports() (called at a distance
compared to where its results are used) to instead be the synchronous
nbd_opt_list() command; the next patch will then focus on making the
API friendlier to async usage by shifting from libnbd malloc'ing the
results to instead having the caller supply a callback function.
Thus, the API signature in this patch is not what we intend to be
final for 1.4, but rather a division to make review of the
transformation easier.

Note that this implementation cannot tell the difference between a
successful return of 0 exports, and a server error.  That will be
fixed in the next patch, with the addition of a completion callback.

I've added the same test for all of C, python, ocaml, and go; but even
though it passes (when new-enough nbdkit is on PATH), I will admit
that the latter two might not be idiomatic.