api: Add nbd_opt_abort and nbd_aio_opt_abort
It is finally time to introduce our first negotiating option command.
With this change, we can now enter NEWSTYLE.START more than once; as
such, it needs to know whether it is the first entry (proceed with
gflags/cflags, TLS, and structured reply) or a later entry (all
nbd_opt_* will cause an IssueCommand event to kick the state machine
out of NEGOTIATING, at which point we want to jump to the sub-state
appropriate for that option).  This is done by setting h->opt_current
prior to entering NEWSTYLE.

We also need a couple of new states to implement NBD_OPT_ABORT; this
was not deemed worth a separate sub-group and new
states-newstyle-opt-abort.c file, so I just inlined it in NEWSTYLE
instead.  One nice benefit of these new states: when NBD_OPT_GO fails
and we are not in opt mode, or when we require TLS but the server does
not support it, we can now hang up gracefully instead of abruptly
disconnecting from a newstyle-fixed server.  This is observable when
using qemu-nbd, in that we now silence the former noise from qemu:

$ truncate --size=512 /tmp/f
$ ./run nbdsh -c '
import errno
try:
  h.connect_systemd_socket_activation(["qemu-nbd", "-f", "raw", "-x", "b",
    "/tmp/f"])
except nbd.Error as e:
  assert e.errnum == errno.ENOENT
'
qemu-nbd: option negotiation failed: Failed to read opts magic: Unexpected end-of-file before all bytes were read

The aio_opt_abort counterpart does not need a completion closure, for
the same reason that aio_disconnect does not have one - the whole
point of this command is a clean shutdown, so you aren't going to rely
on a server reply, and there are no parameters that need cleanup.

This patch adds some testsuite coverage of the new APIs across
multiple languages, although I will admit that my ocaml and golang is
weak enough that the tests there may not be idiomatic, even if they
pass.

Yes, it's a bit awkward that with just this patch, enabling opt_mode
means you HAVE to kill the connection without a data phase; that will
be rectified in the next patch adding nbd_opt_go.  But breaking the
patches up makes review easier.