api: Give aio_opt_go a completion callback
Although NBD_OPT_GO does not need a completion callback (because you
can imply success or failure by the resulting state), it will be a lot
easier to share the code with the upcoming support for NBD_OPT_INFO
(where both success and failure end up back in Negotiating, and thus a
completion callback will be essential for preserving the server
error).  As promised in the previous patch, add in a completion
callback to aio_opt_go.  All future nbd_opt_* commands will also use
a completion callback (nbd_opt_abort remains the only special case
that doesn't need one).

Testing of this will be added later, when NBD_OPT_INFO is added.