interop: Fix VPATH 'make check'
When testing list-exports with nbd-server, we need to point nbd-server
to the config file living in srcdir, not in builddir.  Otherwise, the
testsuite fails when the two are not identical from an in-tree build,
because nbd-server can't find a configuration permitting an export list.

Fixes: c2851cf516
(cherry picked from commit 4230c9b1ebe61fc006fc1f67ae546a5eb9966a9e)